feat: split Cluster sheet/page/file into 5 sections
The single "Cluster" sheet/page/file used to mix everything cluster-wide —
status, options, users, ACL, HA, SDN, pools, mappings — and on large clusters
(25 nodes / 2700+ VMs it grew past 4500 mixed-schema rows in one place,
ungovernable in Excel. It is now broken into five siblings:

* Cluster        — status, options, firewall options, backup jobs, replication,
                   storages, metric servers, hardware mappings
* Cluster Access — users, API tokens, TFA, groups, roles, ACL, domains
* Cluster SDN    — zones, vnets, controllers, IPAMs, subnets
* Cluster HA     — resources, groups, status
* Cluster Pools  — pools with members (VM / CT / storage)

XLSX emits five sheets (ordering preserved via PlaceExact). HTML groups them
under an expandable "Cluster" entry in the sidebar (Overview, Access, SDN, HA,
Pools, then Cluster Log and Cluster Tasks underneath). JSON emits cluster.json,
cluster-access.json, cluster-sdn.json, cluster-ha.json, cluster-pools.json.

Also:
* HTML cover label: "Home" -> "Summary" (filename stays index.html).
* docs(html): rrd-storages.html -> rrd-storage.html in the docs (filename was
  always singular, only the doc was wrong — matches the same fix already done
  in the JSON guide).

Closes #42.
